requirements.txt
pycryptodome==3.19.0
pyotp==2.9.0
qrcode==7.4.2
pillow==10.2.0          # NEW: For handling the QR Code image
customtkinter==5.2.2    # NEW: For the modern Dark UI
argon2-cffi==23.1.0